requirements.txt
awscli==1.24.7
awscli-local==0.20
boto3==1.23.7
botocore==1.26.7
psycopg2-binary==2.9.7
moto==4.2.2